Realized I have some crashes happening. Trying to fix those.
